Refactor code to remove redundancies, implement front end functionality for editing user profile
