+++ client/src/components/Auth/LoginForm.jsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useState} from 'react';
 import {Redirect} from 'react-router-dom';
-import {registerUser, loginUser} from '../../actions/authActions';
+import {loginUser} from '../../actions/authActions';
 import isEmpty from 'is-empty';
 import {REQUEST_SUCCESS} from '../../actions/types';
 import {useStylesForm} from '../../Styles';
@@ -17,18 +17,15 @@ import {
   Typography,
 } from '@material-ui/core';
 
-import FormSubmitMessage from '../layout/FormSubmitMessage';
+import FormSubmitMessage from '../FormSubmitMessage';
 
-function RegisterForm(props) {
+function LoginForm(props) {
   const classes = useStylesForm(themeMain);
 
   const [isLoggedin, setLoggedIn] = useState(props.isLoggedin);
-  const [newUser, setNewUser] = useState({
-    firstName: '',
-    lastName: '',
+  const [user, setUser] = useState({
     email: '',
     password: '',
-    password2: '',
   });
   const [error, setError] = useState({
     errorMessage: '',
@@ -41,7 +38,7 @@ function RegisterForm(props) {
   const handleChange = async (e) => {
     const {name, value} = e.target;
 
-    setNewUser((prevValue) => {
+    setUser((prevValue) => {
       return {
         ...prevValue,
         [name]: value,
@@ -52,33 +49,22 @@ function RegisterForm(props) {
   const handleSubmit = async (e) => {
     e.preventDefault();
 
-    const requestResponse = await registerUser(newUser);
+    const requestResponse = await loginUser(user);
 
     if (requestResponse.authResponseType === REQUEST_SUCCESS) {
-      const requestResponse = await loginUser(newUser);
+      // If login request is successful, clear login form
+      setUser({
+        email: '',
+        password: '',
+      });
 
-      if (requestResponse.authResponseType === REQUEST_SUCCESS) {
-        // If login is successful after registration, clear registration form
-        setNewUser({
-          firstName: '',
-          lastName: '',
-          email: '',
-          password: '',
-          password2: '',
-        });
+      // If login request is successful, clear old errors from state
+      setError({
+        errorMessage: '',
+      });
 
-        // If login is successful after registration, clear old errors
-        setError({
-          errorMessage: '',
-        });
-
-        // Set user as logged in
-        props.handleLoggedIn(true);
-      } else {
-        setError({
-          errorMessage: requestResponse.authResponsePayload,
-        });
-      }
+      // Set user as logged in
+      props.handleLoggedIn(true);
     } else {
       setError({
         errorMessage: requestResponse.authResponsePayload,
@@ -94,40 +80,14 @@ function RegisterForm(props) {
         <CssBaseline />
         <div className={classes.paper}>
           <Typography component='h1' variant='h5'>
-            Start The Oven
+            Sign In Here
           </Typography>
           <form className={classes.form} noValidate>
             <Grid container spacing={2}>
-              <Grid item xs={12} sm={6}>
-                <TextField
-                  onChange={handleChange}
-                  value={newUser.firstName}
-                  variant='outlined'
-                  required
-                  fullWidth
-                  id='firstName'
-                  label='First Name'
-                  name='firstName'
-                  autoComplete='fname'
-                />
-              </Grid>
-              <Grid item xs={12} sm={6}>
-                <TextField
-                  onChange={handleChange}
-                  value={newUser.lastName}
-                  variant='outlined'
-                  required
-                  fullWidth
-                  id='lastName'
-                  label='Last Name'
-                  name='lastName'
-                  autoComplete='lname'
-                />
-              </Grid>
               <Grid item xs={12}>
                 <TextField
                   onChange={handleChange}
-                  value={newUser.email}
+                  value={user.email}
                   variant='outlined'
                   required
                   fullWidth
@@ -140,7 +100,7 @@ function RegisterForm(props) {
               <Grid item xs={12}>
                 <TextField
                   onChange={handleChange}
-                  value={newUser.password}
+                  value={user.password}
                   variant='outlined'
                   required
                   fullWidth
@@ -148,21 +108,7 @@ function RegisterForm(props) {
                   label='Password'
                   type='password'
                   id='password'
-                  autoComplete='current-password'
-                />
-              </Grid>
-              <Grid item xs={12}>
-                <TextField
-                  onChange={handleChange}
-                  value={newUser.password2}
-                  variant='outlined'
-                  required
-                  fullWidth
-                  name='password2'
-                  label='Confirm password'
-                  type='password'
-                  id='password2'
-                  autoComplete='confirm-password'
+                  autoComplete='password'
                 />
               </Grid>
               <Grid item xs={12}>
@@ -174,14 +120,19 @@ function RegisterForm(props) {
                   color='primary'
                   className={classes.submit}
                 >
-                  Sign Up
+                  Login
                 </Button>
               </Grid>
             </Grid>
-            <Grid container justify='flex-end' className={classes.form}>
+            <Grid container justify='space-between' className={classes.form}>
               <Grid item>
-                <Link href='/login' variant='body2'>
-                  Already have an account? Sign in
+                <Link href='/forgotpassword' variant='body2'>
+                  Forgot Password?
+                </Link>
+              </Grid>
+              <Grid item>
+                <Link href='/' variant='body2'>
+                  Don't have an account? Register
                 </Link>
               </Grid>
             </Grid>
@@ -197,4 +148,4 @@ function RegisterForm(props) {
   );
 }
 
-export default RegisterForm;
+export default LoginForm;
