+++ client/src/components/FormSubmitMessage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {useStylesFormSubmitMsg} from '../../Styles';
+import {useStylesFormSubmitMsg} from '../Styles';
 
 import {Typography} from '@material-ui/core';
